requirements.txt
tensorflow>=2.10.0
opencv-python>=4.7.0
pillow>=9.4.0
matplotlib>=3.6.0
numpy>=1.23.0